Minor refactoring for coherency
