+++ src/__test__/redux/actions/embeddings/loadEmbedding.test.js
@@ -2,7 +2,7 @@ import configureStore from 'redux-mock-store';
 import thunk from 'redux-thunk';
 import { loadEmbedding } from '../../../../redux/actions/embedding';
 import { initialEmbeddingState } from '../../../../redux/reducers/embeddings/initialState';
-import generateExperimentSettingsMock, { initialPipelineState } from '../../../test-utils/experimentSettings.mock';
+import generateExperimentSettingsMock from '../../../test-utils/experimentSettings.mock';
 
 import {
   EMBEDDINGS_ERROR,
@@ -23,11 +23,21 @@ const embeddingType = 'umap';
 
 const initialExperimentState = generateExperimentSettingsMock([]);
 
+const initialPipelineState = {
+  startDate: null,
+  endDate: null,
+  status: null,
+  completedSteps: [],
+};
+
 describe('loadEmbedding action', () => {
   const experimentId = '1234';
   const experimentSettings = {
     ...initialExperimentState,
-    backendStatus: {
+  };
+
+  const backendStatus = {
+    [experimentId]: {
       status: {
         pipeline: {
           ...initialPipelineState,
@@ -44,6 +54,7 @@ describe('loadEmbedding action', () => {
   it('Dispatches if not loaded', async () => {
     const store = mockStore(
       {
+        backendStatus,
         experimentSettings,
         embeddings: {},
       },
@@ -56,6 +67,7 @@ describe('loadEmbedding action', () => {
   it('Does not dispatch if embedding is already loaded', async () => {
     const store = mockStore(
       {
+        backendStatus,
         experimentSettings,
         embeddings:
         {
@@ -78,6 +90,7 @@ describe('loadEmbedding action', () => {
   it('Does not dispatch on a loading embedding', async () => {
     const store = mockStore(
       {
+        backendStatus,
         experimentSettings,
         embeddings:
           { [embeddingType]: { ...initialEmbeddingState, loading: true } },
@@ -107,6 +120,7 @@ describe('loadEmbedding action', () => {
 
     const store = mockStore(
       {
+        backendStatus,
         embeddings: {},
         experimentSettings,
       },
@@ -145,6 +159,7 @@ describe('loadEmbedding action', () => {
 
     const store = mockStore(
       {
+        backendStatus,
         embeddings:
           { [embeddingType]: { ...initialEmbeddingState, error: true, loading: false } },
         experimentSettings,
@@ -168,6 +183,7 @@ describe('loadEmbedding action', () => {
   it('Dispatches error action on unsuccessful loading', async () => {
     const store = mockStore(
       {
+        backendStatus,
         embeddings: {},
         experimentSettings,
       },
@@ -192,6 +208,7 @@ describe('loadEmbedding action', () => {
   it('Does not return anything while waiting for config data to load', async () => {
     const store = mockStore(
       {
+        backendStatus,
         embeddings: {},
         experimentSettings: {
           ...experimentSettings,
@@ -209,13 +226,16 @@ describe('loadEmbedding action', () => {
   it('Dispatches error if pipeline has not been run', async () => {
     const store = mockStore(
       {
+        backendStatus: {
+          ...backendStatus,
+          1234: {
+            ...backendStatus['1234'],
+            status: {},
+          },
+        },
         embeddings: {},
         experimentSettings: {
           ...experimentSettings,
-          backendStatus: {
-            ...experimentSettings.backendStatus,
-            status: {},
-          },
         },
       },
     );
@@ -246,6 +266,7 @@ describe('loadEmbedding action', () => {
 
     const store = mockStore(
       {
+        backendStatus,
         embeddings:
           { [embeddingType]: { ...initialEmbeddingState, error: false, loading: false } },
         experimentSettings,
