+++ src/__test__/test-utils/experimentSettings.mock.js
@@ -1,7 +1,4 @@
-import initialState, {
-  initialPipelineState,
-  initialWorkerState,
-} from '../../redux/reducers/experimentSettings/initialState';
+import initialState from '../../redux/reducers/experimentSettings/initialState';
 
 const sampleifiedConfig = (sampleIds, configToReplicate) => {
   const result = sampleIds.reduce(
@@ -135,5 +132,3 @@ const generateExperimentSettingsMock = (sampleIds) => ({
 });
 
 export default generateExperimentSettingsMock;
-
-export { initialPipelineState, initialWorkerState };
