+++ src/redux/reducers/experimentSettings/initialState.js
@@ -1,17 +1,3 @@
-const initialPipelineState = {
-  startDate: null,
-  endDate: null,
-  status: null,
-  completedSteps: [],
-};
-
-const initialWorkerState = {
-  status: null,
-  started: false,
-  ready: null,
-  restartCount: 0,
-};
-
 const initialState = {
   info: {
     experimentId: null,
@@ -30,4 +16,3 @@ const initialState = {
 };
 
 export default initialState;
-export { initialPipelineState, initialWorkerState };
